requirements.txt
torch
torchvision
diffusers
transformers
accelerate
opencv-python
numpy
scikit-learn
umap-learn
matplotlib
tqdm
Pillow
